backend/requirements/base.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Content & Media
Pillow==10.1.0
django-ckeditor==6.7.0
django-taggit==4.0.0

# Web Crawling
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Text Processing & Similarity
nltk==3.8.1
scikit-learn==1.3.2
python-dateutil==2.8.2

# Utilities
python-slugify==8.0.1
bleach==6.1.0
python-magic==0.4.27

# API Documentation
drf-spectacular==0.26.5
